feat: move desktopApp & fix & extended packaging
