+++ .github/workflows/nightly.yml
@@ -14,10 +14,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - name: Set up JDK 17
+      - name: Set up JDK 21
         uses: actions/setup-java@v4
         with:
-          java-version: '17'
+          java-version: '21'
           distribution: 'temurin'
           cache: gradle
           
@@ -37,7 +37,7 @@ jobs:
            SIGNING_KEY_ALIAS: ${{ secrets.SIGNING_KEY_ALIAS }}
            SIGNING_KEY_PASSWORD: ${{ secrets.SIGNING_KEY_PASSWORD }}
            SIGNING_STORE_PASSWORD: ${{ secrets.SIGNING_STORE_PASSWORD }}
-        run: ./gradlew assembleNightly
+        run: ./gradlew :androidApp:assembleNightly
 
       - name: Set time
         run: |
@@ -48,4 +48,4 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
           tag: "nightly"
           files: |
-            ./composeApp/build/outputs/apk/nightly/composeApp-nightly.apk
+            ./androidApp/build/outputs/apk/nightly/androidApp-nightly.apk
