+++ build.gradle.kts
@@ -4,6 +4,7 @@ import java.util.Properties
 
 plugins {
     alias(libs.plugins.multiplatform).apply(false)
+    alias(libs.plugins.kotlin.jvm).apply(false)
     alias(libs.plugins.compose.compiler).apply(false)
     alias(libs.plugins.compose).apply(false)
     alias(libs.plugins.android.application).apply(false)
@@ -31,3 +32,7 @@ val kitshnAndroidPackageName by extra("de.kitshn.android")
 val kitshnDesktopPackageName by extra("kitshn")
 
 val kitshnIsBeta by extra(false)
+
+tasks.register("printKitshnVersionName") {
+    doLast { println(kitshnVersionName) }
+}
