+++ desktopApp/desktopAppIcons/MacosIcon.icns
