+++ desktopApp/packaging/appimage/AppRun
@@ -1,0 +1,3 @@
+#!/bin/sh
+HERE="$(dirname "$(readlink -f "$0")")"
+exec "$HERE/bin/kitshn" "$@"
