+++ desktopApp/src/main/kotlin/main.kt
@@ -1,12 +1,12 @@
-import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import de.kitshn.App
-import dev.datlag.kcef.KCEF
+import de.kitshn.disposeKcefBlocking
 
 fun main() = application {
     Window(
@@ -19,7 +19,7 @@ fun main() = application {
 
     DisposableEffect(Unit) {
         onDispose {
-            KCEF.disposeBlocking()
+            disposeKcefBlocking()
         }
     }
 }
