+++ settings.gradle.kts
@@ -33,3 +33,4 @@ dependencyResolutionManagement {
 }
 include(":shared")
 include(":androidApp")
+include(":desktopApp")
