+++ shared/src/jvmMain/kotlin/de/kitshn/KcefLifecycle.jvm.kt
@@ -1,0 +1,5 @@
+package de.kitshn
+
+import dev.datlag.kcef.KCEF
+
+fun disposeKcefBlocking() = KCEF.disposeBlocking()
